修复边界检查并精简未使用的 gd 字段
限制 USB PD PDO 数量为 7 个对象，并取消 RX 缓冲区下标环绕写入，避免异常长包写穿 pd_rx_buff。

限制 AFC/SCP 多字节读写长度，同时考虑协议 payload 上限和 SCP_REG 剩余空间，并用裁剪后的长度生成响应。

删除未调用的电池欠压 forbid 检查函数，以及 bat_uv_forbid_flag 相关配置、声明、UI 显示和日志引用，保留现有过压 forbid 流程。

将 forbid bypass 和 SOC 睡眠备份状态从 gd_t 挪到普通全局变量，降低固定 gd RAM 区压力，同时保持原有调用逻辑。

修正 app.c 中电流采样排序的临时变量类型，避免 16 位 ADC 值经过 uint8_t 截断。

同步 Eclipse 工程名为 init 构建目标。

+++ app/app.c
@@ -165,7 +165,7 @@ void apl_task_event_handler(uint32_t event)
 		        {
 		            if (u16IsnsTmp[j] < u16IsnsTmp[j + 1])
 		            {
-		            	uint8_t temp = u16IsnsTmp[j];
+		            uint16_t temp = u16IsnsTmp[j];
 		            	u16IsnsTmp[j] = u16IsnsTmp[j + 1];
 		            	u16IsnsTmp[j + 1] = temp;
 		            }
